pharmassist/requirements.txt
alabaster==0.7.7
amqp==1.4.9
anyjson==0.3.3
argh==0.26.1
Babel==2.2.0
billiard==3.3.0.22
binaryornot==0.4.0
celery==3.1.19
chardet==2.3.0
click==6.2
cookiecutter==1.3.0
cov-core==1.15.0
coverage==4.0.3
decorator==4.0.6
defusedxml==0.4.1
Django==1.9
django-allauth==0.24.1
django-autoslug==1.9.3
django-braces==1.8.1
django-coverage-plugin==1.1
django-crispy-forms==1.5.2
django-debug-toolbar==1.4
django-environ==0.4.0
django-extensions==1.5.9
django-filter==0.12.0
django-floppyforms==1.6.0
django-model-utils==2.3.1
django-redis==4.3.0
django-secure==1.0.1
django-test-plus==1.0.11
djangorestframework==3.3.2
docutils==0.12
factory-boy==2.6.0
fake-factory==0.5.3
flake8==2.2.2
future==0.15.2
ghp-import==0.4.1
ipdb==0.8.1
ipython==4.0.1
ipython-genutils==0.1.0
Jinja2==2.8
kombu==3.0.33
Markdown==2.4.1
MarkupSafe==0.23
mccabe==0.3.1
mkdocs==0.11.1
oauthlib==1.0.3
path.py==8.1.2
pathtools==0.1.2
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.5
Pillow==3.0.0
psycopg2==2.6.1
ptyprocess==0.5
py==1.4.31
pyflakes==1.0.0
Pygments==2.0.2
pytest==2.5.2
pytest-cov==1.6
pytest-django==2.6
python3-openid==3.0.9
pytz==2015.7
PyYAML==3.11
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.0
ruamel.base==1.0.0
ruamel.yaml==0.10.13
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.3.3
sphinx-rtd-theme==0.1.9
sqlparse==0.1.18
traitlets==4.0.0
unicode-slugify==0.1.3
Unidecode==0.4.18
watchdog==0.8.3
Werkzeug==0.11.2
wheel==0.24.0
whichcraft==0.1.1
